chore : Seperate entities
