+++ src/domain/entities/github_user.py
@@ -1,11 +1,4 @@
-class Group:
-    def __init__(self, id: int, name: str, category: str):
-        self.id = id
-        self.name = name
-        self.category = category
-
-    def change_category(self, new_caregory: str):
-        self.category = new_caregory
+from src.domain.entities.group import Group
 
 
 class GithubUser:
