+++ src/domain/entities/group.py
@@ -1,0 +1,8 @@
+class Group:
+    def __init__(self, id: int, name: str, category: str):
+        self.id = id
+        self.name = name
+        self.category = category
+
+    def change_category(self, new_caregory: str):
+        self.category = new_caregory
